+++ docs/guides/changelog.md
@@ -9,6 +9,7 @@ description: |-
 
 ## 0.2.4 (unreleased)
 
+- Add `redeploy` attribute to `catalystcenter_deploy_template`
 - Add `catalystcenter_fabric_l3_handoff_ip_transits` resource and data source
 - Add `catalystcenter_update_device_management_address` resource
 - Fix issue with terraform not detecting delta in `catalystcenter_device` resource, [link](https://github.com/CiscoDevNet/terraform-provider-catalystcenter/issues/200)